third_party/WebKit/LayoutTests/csspaint/style-background-image-expected.txt

CONSOLE MESSAGE: line 14: --bar: [null]
CONSOLE MESSAGE: line 14: --foo: [CSSStyleValue= bar]
CONSOLE MESSAGE: line 14: align-items: [CSSKeywordValue=normal]
CONSOLE MESSAGE: line 14: border-radius: [CSSStyleValue=2px]
This tests the style information in the paint callback.

See the devtools console for test output. The console should log:

--bar: [null]
--foo: [CSSStyleValue= bar]
align-items: [CSSKeywordValue=normal]
border-radius: [CSSStyleValue=2px]
